requirements.txt
ackermann-msgs==2.0.2
action-msgs==1.2.2
actionlib-msgs==4.9.1
actuator-msgs==0.0.1
addict==2.4.0
ament-cmake-test==1.3.13
ament-copyright==0.12.15
ament-cppcheck==0.12.15
ament-cpplint==0.12.15
ament-flake8==0.12.15
ament-index-python==1.4.1
ament-lint==0.12.15
ament-lint-cmake==0.12.15
ament-package==0.14.2
ament-pep257==0.12.15
ament-uncrustify==0.12.15
ament-xmllint==0.12.15
angles==1.15.0
apriltag-detection==0.0.1
apriltag-msgs==2.0.1
aruco-msgs==5.0.5
aruco-opencv-msgs==2.4.1
asttokens==3.0.1
attrs==26.1.0
blinker==1.9.0
bond==4.1.2
builtin-interfaces==1.2.2
camera-calibration==3.0.9
cartographer-ros-msgs==2.0.9002
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
comm==0.2.3
composition-interfaces==1.2.2
ConfigArgParse==1.7.5
contourpy==1.3.2
control-msgs==4.8.0
controller-manager==2.53.1
controller-manager-msgs==2.53.1
cv-bridge==3.2.1
cycler==0.12.1
dash==4.0.0
decorator==5.2.1
depthai-ros-driver==2.12.2
depthai-ros-msgs==2.12.2
diagnostic-aggregator==4.0.6
diagnostic-msgs==4.9.1
diagnostic-updater==4.0.6
domain-coordinator==0.10.0
dwb-msgs==1.1.20
dynamixel-sdk==3.8.4
dynamixel-sdk-custom-interfaces==3.8.4
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
ffmpeg-image-transport-msgs==1.1.2
Flask==3.1.3
fonttools==4.62.1
foxglove-msgs==3.2.4
gazebo-msgs==3.9.0
generate-parameter-library-py==0.6.0
geometry-msgs==4.9.1
gps-msgs==2.1.2
graph-msgs==0.2.0
grid-map-msgs==2.0.1
idna==3.11
image-geometry==3.2.1
importlib_metadata==8.7.1
interactive-markers==2.3.2
ipython==8.38.0
ipywidgets==8.1.8
irobot-create-common-bringup==2.1.0
irobot-create-msgs==2.1.0
isaac-tb3-bringup==0.0.0
itsdangerous==2.2.0
jedi==0.19.2
jetbot-camera==0.0.0
jetbot-nav2==0.0.0
jetbot-semantic==0.1.0
jetbot-teleoperate==0.0.0
jetbot-vision-nav==0.0.1
Jinja2==3.1.6
joblib==1.5.3
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.5.0
laser-geometry==2.4.1
launch==1.0.14
launch-param-builder==0.1.1
launch-ros==0.19.13
launch-testing==1.0.14
launch-testing-ros==0.19.13
launch-xml==1.0.14
launch-yaml==1.0.14
lifecycle-msgs==1.2.2
linkattacher-msgs==0.0.0
map-msgs==2.1.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
message-filters==4.3.14
# Editable install with no version control (morefusion==1.0.0)
-e /home/roboticslab/morefusion_ws/build/morefusion
moveit==0.1.3
moveit-configs-utils==2.5.9
moveit-msgs==2.2.1
moveit-task-constructor-msgs==0.1.3
narwhals==2.18.0
nav-2d-msgs==1.1.20
nav-msgs==4.9.1
nav2-common==1.1.20
nav2-msgs==1.1.20
nav2-simple-commander==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
object-recognition-msgs==2.0.0
octomap-msgs==2.0.1
open3d==0.19.0
# Editable install with no version control (opencv==0.0.0)
-e /home/roboticslab/ur_ws/build/opencv
opencv-python==4.13.0.92
osrf-pycommon==2.1.6
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pandas==2.3.3
parso==0.8.6
pcl-msgs==1.0.0
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.2
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-qt-binding==1.1.3
pytz==2025.2
PyYAML==6.0.3
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
rcl-interfaces==1.2.2
rclpy==3.3.20
rcutils==5.1.8
realsense2-camera-msgs==4.56.4
referencing==0.37.0
requests==2.32.5
resource-retriever==3.1.3
retrying==1.4.2
rmw-dds-common==1.6.0
robot-upstart==1.0.4
ros-gz-interfaces==0.244.22
ros-ign-interfaces==0.244.22
ros2-controllers-test-nodes==2.52.0
ros2action==0.18.17
ros2bag==0.15.16
ros2cli==0.18.17
ros2component==0.18.17
ros2controlcli==2.53.1
ros2doctor==0.18.17
ros2interface==0.18.17
ros2launch==0.19.13
ros2lifecycle==0.18.17
ros2multicast==0.18.17
ros2node==0.18.17
ros2param==0.18.17
ros2pkg==0.18.17
ros2plugin==5.1.3
ros2run==0.18.17
ros2service==0.18.17
ros2topic==0.18.17
rosbag2-interfaces==0.15.16
rosbag2-py==0.15.16
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.8
rosidl-cli==3.1.8
rosidl-cmake==3.1.8
rosidl-generator-c==3.1.8
rosidl-generator-cpp==3.1.8
rosidl-generator-py==0.14.6
rosidl-parser==3.1.8
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.4
rosidl-typesupport-fastrtps-cpp==2.2.4
rosidl-typesupport-introspection-c==3.1.8
rosidl-typesupport-introspection-cpp==3.1.8
rpds-py==0.30.0
rpyutils==0.2.2
rqt==1.1.9
rqt-graph==1.3.2
rqt-gui==1.1.9
rqt-gui-py==1.1.9
rqt-py-common==1.1.9
rqt-robot-monitor==1.0.6
rtabmap-msgs==0.22.1
rtabmap-python==0.1.0
scikit-learn==1.7.2
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1739790642651/work/dist/scipy-1.15.2-cp310-cp310-linux_x86_64.whl#sha256=9e52bad6c3294d1a5b04a13632118ca2157130603c6c018c2d710162b223b27e
scripts==3.9.0
sensor-msgs==4.9.1
sensor-msgs-py==4.9.1
shape-msgs==4.9.1
simple-term-menu-vendor==1.5.7
six==1.17.0
slam-toolbox==2.6.10
smclib==4.1.2
srdfdom==2.0.7
sros2==0.10.9
stack-data==0.6.3
statistics-msgs==1.2.2
std-msgs==4.9.1
std-srvs==4.9.1
stereo-msgs==4.9.1
teleop-twist-keyboard==2.4.1
tf-transformations==1.1.1
tf2-geometry-msgs==0.25.19
tf2-kdl==0.25.19
tf2-msgs==0.25.19
tf2-py==0.25.19
tf2-ros-py==0.25.19
tf2-sensor-msgs==0.25.19
tf2-tools==0.25.19
theora-image-transport==2.5.4
threadpoolctl==3.6.0
topic-tools==1.1.1
topic-tools-interfaces==1.1.1
tqdm==4.67.3
traitlets==5.14.3
trajectory-msgs==4.9.1
transforms3d==0.4.2
turtlebot3-example==2.3.6
turtlebot3-msgs==2.4.0
turtlebot3-teleop==2.3.6
turtlebot4-diagnostics==1.0.3
turtlebot4-msgs==1.0.5
turtlebot4-navigation==1.0.5
turtlebot4-python-tutorials==1.0.1
turtlebot4-setup==1.0.6
turtlebot4-tests==1.0.3
typing_extensions==4.15.0
tzdata==2025.3
unique-identifier-msgs==2.2.1
ur-dashboard-msgs==2.12.0
ur-msgs==2.3.0
ur-robot-driver==2.12.0
ur-sim==0.0.0
urdfdom-py==1.2.1
urllib3==2.6.3
velodyne-msgs==2.5.1
vision-msgs==4.1.1
visualization-msgs==4.9.1
wcwidth==0.6.0
Werkzeug==3.1.6
widgetsnbextension==4.0.15
xacro==2.1.1
zipp==3.23.0
